+++ src/grpc-server/src/Server.php
@@ -12,6 +12,29 @@ declare(strict_types=1);
 
 namespace Hyperf\GrpcServer;
 
-interface Server
+use Hyperf\Contract\ConfigInterface;
+use Hyperf\Dispatcher\HttpDispatcher;
+use Hyperf\GrpcServer\Exception\Handler\GrpcExceptionHandler;
+use Hyperf\HttpServer\Server as HttpServer;
+use Psr\Container\ContainerInterface;
+
+class Server extends HttpServer
 {
+    public function __construct(ContainerInterface $container)
+    {
+        $this->container = $container;
+        $this->dispatcher = $container->get(HttpDispatcher::class);
+    }
+
+    public function initCoreMiddleware(string $serverName): void
+    {
+        $this->serverName = $serverName;
+        $this->coreMiddleware = new CoreMiddleware($this->container, $serverName);
+
+        $config = $this->container->get(ConfigInterface::class);
+        $this->middlewares = $config->get('middlewares.' . $serverName, []);
+        $this->exceptionHandlers = $config->get('exceptions.handler.' . $serverName, [
+            GrpcExceptionHandler::class,
+        ]);
+    }
 }