+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Fixed
 
+- [#235](https://github.com/hyperf-cloud/hyperf/pull/235) Added default exception handler for `grpc-server` and optimized code.
 - [#240](https://github.com/hyperf-cloud/hyperf/pull/240) Fixed OnPipeMessage event will be dispatch by another listener.
 
 # v1.0.5 - 2019-07-07